+++ src/main/java/team/hanaro/hanamate/domain/Login/RefreshTokenRepository.java
@@ -1,0 +1,8 @@
+package team.hanaro.hanamate.domain.Login;
+
+import java.util.Optional;
+
+public interface RefreshTokenRepository {
+    Optional<RefreshToken> findByAccountLoginId(String loginId);
+
+}
